third_party/blink/web_tests/platform/generic/http/tests/security/subresourceIntegrity/subresource-integrity-invalid-integrity-console-messages-expected.txt

CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256;B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
ALERT: PASS
CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256:B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
ALERT: PASS
CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256-(B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg=)'). The digest must be a valid, base64-encoded value.
ALERT: PASS
